Fit last render window of rows into the display window when scrolling
Introduced method of adding margin to the scroll spacer and table data
container when the final render group of rows is scrolled to.

Leaving (commented out) alternative method for handling specific case
of fitting final row into an unevenly-heighted display window when all
rows are default height.

Will likely remove commented code solution if the code that is
currently handling this case and case when rows are larger than default
height proves effective.