Fix resize columns problems, fix scroll to bottom issue, made a pass at styling UI
Resizing columns UI regressed with various changes made to unrelated
functionality, so this was given a once over. Reordering columns will
need the same treatment.

Issue discovered when the scrollToRow method is called for a row in the
last DOM row window and these rows are taller than default. The bottom
margin was not being applied like it is when naturally scrolling to the
same position, so that functionality was broken out to be used in
various other senarios when it's needed.

Rearranged the order in which events are defined in _create() purely
for organizational purposes.

Made a styling pass on the UI. There will be more work to do, but this
is intended just to get the ball rolling.