Change column reorder to use a new handle icon rather than click and hold
Reordered many event listener definitions in _change to prepare for
eventual refactoring of this function.

Major change in the way reordering columns are processed, both UX and
under the covers.

_renderTableHeader/Rows now maintain the scroll state they were
originally in after processing is complete.

Broke the web worker section of the _sortTable function into a seperate
helper function.