Issue #30 introduce fit-to-available-space, calculated column width, fix scrolling to bottom/right bugs
Refactor order of header rendering to allow the truly available
horizontal space to be calculated in order to allow columns to fill
space evenly, automatically.

There was a bug where you couldn't scroll all the way to the right to
see the full, last column in certain cases, especially if the header
control static columns were showing.

There was a bug scrolling down to the end of the table where
padding-bottom was being incorrectly applied and the current row was
being misidentified, throwing errors.