#64 change behavior of sort order when switching between columns
Now in order to get a column to sort in descending order via the UI,
you must click on a column header twice. Whenever you click on a column
header that isn't currently the sorted column, the sort order will
start off with ascending.

When reseting the tableData, all sort orders will be reset.